refact: EditCardActivity 시작하는 방식 변경
